requirements.txt
habitat-sim>=0.2.3
opencv-python>=4.5.0
requests>=2.28.0
numpy>=1.21.0
pillow>=8.3.0
torch>=1.12.0
transformers>=4.20.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
tensorboard>=2.8.0
pyyaml>=6.0